add learning and growth process

+++ FormChao/Save/SaveParameters.cs
@@ -5,7 +5,7 @@ namespace BebooGarden.Save;
 
 public class SaveParameters
 {
-  public SaveParameters(string? language, float volume, string bebooName, int age, float energy, int happiness,
+  public SaveParameters(string? language, float volume, string bebooName, float age, float energy, int happiness,
       DateTime lastPayed, Flags flags, string playerName, SortedDictionary<FruitSpecies, int> fruitsBasket,
       int remainingFruits, List<Item> inventory, int tickets, List<Item> mapItems, List<string> unlockedRolls, string favoredColor)
   {
@@ -44,7 +44,7 @@ public class SaveParameters
   public float Volume { get; set; }
   public string? Language { get; set; }
   public string BebooName { get; set; }
-  public int Age { get; set; }
+  public float Age { get; set; }
   public DateTime LastPlayed { get; set; }
   public float Energy { get; set; }
   public int Happiness { get; set; }
